% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screen.R
\name{assignProfiles}
\alias{assignProfiles}
\title{Merge sample list information to a profile container}
\usage{
assignProfiles(profiles, sampleList, zerofill = FALSE)
}
\arguments{
\item{profiles}{EnviMass profile container or list of containers}

\item{sampleList}{A sample list with attached sampleIDs, as obtained from \link{assignSamples}.}

\item{zerofill}{If \code{zerofill} is active, a new zero-intensity result is generated for every possible profile-sample combination
(otherwise many results are "missing").}
}
\value{
The updated profile container(s).
}
\description{
Adds sample list columns (e.g. timepoint, species, etc as specified in the sample list) to the \code{peaks} table in the \code{profiles} container,
such that it can later easily be sorted, filtered etc.
}
\author{
stravsmi
}
