% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchPick.R
\name{batchPick}
\alias{batchPick}
\alias{batchPickDIA}
\alias{comprehensiveBatchPick}
\title{Batch pick raw mzXML files}
\usage{
batchPick(
  files,
  outputDir,
  polarity = c("+", "-"),
  writeData = TRUE,
  writeList = TRUE,
  settings = getOption("RMassScreening")$enviPick,
  multicore = FALSE,
  log = "logCluster.txt"
)

batchPickDIA(
  files,
  outputDir,
  writeData = TRUE,
  writeList = TRUE,
  settings = getOption("RMassScreening")$enviPick,
  multicore = FALSE,
  log = "logCluster.txt"
)

comprehensiveBatchPick(...)
}
\arguments{
\item{files}{Array of paths to the raw files.}

\item{outputDir}{Folder to write picked file results to.}

\item{polarity}{For \code{batchPick}, \code{+}or \code{-}. Selects the polarity to pick. On one hand, this is needed to handle polarity switching files correctly;
on the other hand, it also determines the file names to use downstream.}

\item{writeData}{If \code{TRUE}, the picked file object (the \code{MSlist} from enviPick) is saved to disk}

\item{writeList}{If \code{TRUE}, the CSV picked peak table is saved to disk}

\item{settings}{Settings to use for enviPick. Read the settings file template for descriptions, or the enviPick documentation accordingly.
By default uses the settings as loaded.}

\item{multicore}{\code{FALSE} for single core processing, or an integer number for the number of cores to use.}

\item{log}{Logfile output for multicore processing (irrelevant for single core).}
}
\value{
For \code{batchPickDIA}, returns a character vector of all scan types picked (the filename patterns for use in \code{fillProfiles} etc.).
}
\description{
Peak picking using mostly enviPick. The function handles polarity switching raw files and parallel picking on multiple cores.
\code{batchPickDIA} additionally handles DIA files, including polarity switching DIA files, and picks every DIA scan separately.
}
\details{
Warning: \code{batchPick} doesn't actually use the \code{settings} at all! only batchPickDIA does.

\code{batchPick} picks either positive or negative mode data as specified. For a raw file \code{RAWFILE.mzXML}, the picked files (\code{writeData} output) have names
\code{RAWFILE.mzXML.MSlist.pos.RData} for positive mode, or \code{.neg.RData} for negative mode respectively.  For \code{writeList}, the files are \code{.csv} instead.

For \code{batchPickDIA}, the file names follow the pattern \code{RAWFILE.mzXML.MSlist.LEVEL-POLARITY-CENTER-WIDTH.RData}. For the MS1 scans,
this is e.g. \verb{1-pos--} because there is no isolation center or isolation width. For the MS2 scans, a positive scan with center 400 and isolation width
200 would have \code{2-pos-400-200} in the name.

For legacy reasons, \code{batchPick} produces the objects \code{MSlist.pos} and \code{MSlist.neg} respectively, whereas \code{batchPickDIA}
only makes \code{MSlist} regardless of polarity.

\code{batchPick} will soon be deprecated (or rather \code{batchPickDIA} will take over the name.)

If \code{multicore} is used, the processing messages ("output") are piped to the \code{logfile}. For single core, the output is displayed on screen.
}
\section{Functions}{
\itemize{
\item \code{batchPickDIA}: 

\item \code{comprehensiveBatchPick}: 
}}

\note{
\code{batchPickDIA} is the new name for \code{comprehensiveBatchPick}, which should not be used anymore.
}
\author{
stravsmi
}
