% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/masses.R
\name{combineReactions.formula}
\alias{combineReactions.formula}
\title{Combine reactions using molecular formula, not mass.}
\usage{
combineReactions.formula(substances, reactions, sep = " ", omit.name = "")
}
\description{
Expects two data frames with columns mass, name.
Note that the "reactions" dataframe will have dash-separated names,
such as "oh-deme"
}

