% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/masses.R
\name{consolidateReactions}
\alias{consolidateReactions}
\title{Merge reactions with identical mass differences}
\usage{
consolidateReactions(reactions, eps = 2e-04, sep = "-")
}
\arguments{
\item{reactions}{Data frame with \code{name} and \code{mass}.}

\item{eps}{Merging margin, absolute}

\item{sep}{Reaction separator token to filter/count for priority}
}
\description{
Merges reactions with identical mass difference within \code{eps} to avoid creating duplicate reaction products.
Of multiple reactions, the "simplest" one will be kept, i.e. if one is already combined (contains a separator as specified
by \code{sep} and the other is not, the "shorter" path will be kept.
}
\author{
stravsmi
}
