% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/masses.R
\name{convertFormulas}
\alias{convertFormulas}
\title{Convert formulas to masses}
\usage{
convertFormulas(substances)
}
\arguments{
\item{substances}{a data.frame with column \code{formula} and existing column \code{mass}.}
}
\value{
Data frame with overwritten column \code{mass}.
}
\description{
Calculates the monoisotopic exact masses from formulas.
}
\author{
stravsmi
}
