% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profiles.R
\name{cutSingleProfiles}
\alias{cutSingleProfiles}
\title{Cut single profiles}
\usage{
cutSingleProfiles(profiles, minPeaks = 1)
}
\arguments{
\item{profiles}{A profile container or list of profile containers}

\item{minPeaks}{Minimum peak count required}
}
\value{
Filtered profiles.
}
\description{
Removes profiles which do not have at least a specified number of peaks. Typically the limit is 1, i.e.
requiring a profile to consist of at least two peaks.
}
\note{
Check: Should the `profile_ID` `index_prof` also be rewritten?
}
\author{
stravsmi
}
