% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchPick.R
\name{.call.fun.settings}
\alias{.call.fun.settings}
\title{Call function with specified settings}
\usage{
.call.fun.settings(fun, settings, ...)
}
\arguments{
\item{fun}{(character) Function name to be called}

\item{settings}{Settings (options) list}

\item{...}{Additional settings to substitute into \code{fun} that are not in \code{settings}}
}
\description{
Wrapper for formals() and do.call() which finds the arguments of \code{fun} and substitutes them
with the data in \code{settings} (but \code{settings} can have many more settings stored)
Arguments in \code{...} take precedence over those in \code{settings}
}
\examples{
#nope
}
