% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchPick.R
\name{.getScantypeSettings}
\alias{.getScantypeSettings}
\title{Get settings per scantype}
\usage{
.getScantypeSettings(allTypes, settings = getOption("RMassScreening")$enviPick)
}
\arguments{
\item{allTypes}{All scan types to be processed}

\item{settings}{enviPick settings in the format used in the enviPick listslot in settings.ini

This processes the settings such that the global settings are taken for all scan types,
and then for every specified scantype the global settings are overwritten with specific
settings if they exist (e.g. if a scantype has a specific "maxint" defined).}
}
\value{

}
\description{
Get settings per scantype
}
