% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profiles.R
\name{fillProfiles}
\alias{fillProfiles}
\title{Fill enviMass profiles from picked files}
\usage{
fillProfiles(dataDir, files, polarity = "+", pattern = NA,
  rtrange = c(-Inf, Inf))
}
\arguments{
\item{dataDir}{Folder containing the enviPick files in .MSlist format. May be either MSlist.pos and MSlist.neg files, or files with an arbitrary pattern.
Typically, filenames should correspond to the names built by \link{batchPick} or \link{comprehensiveBatchPick}, respectively:
* If only MS1 data was peak-picked (\link{batchPick}), raw file \code{FILENAME.mzXML} should give filenames \code{FILENAME.mzXML.MSlist.pos.RData}
(or \code{neg}, respectively) if only MS1 were picked
* If DIA data was picked (\code{\link{comprehensiveBatchPick}}) a raw file \code{FILENAME.mzXML} should give filenames:
* \code{FILENAME.mzXML.MSlist.1-pos--} for the MS1 positive
* e.g. \code{FILENAME.mzXML.MSlist.2-neg-200-100} for MS2-DIA, negative, center mass 200, isolation width 100 etc.
However other patterns work too, since \code{pattern} specifies which file pattern is searched.}

\item{files}{The filenames of the raw files, with or without full path. NOT the filenames or path of the enviPick results.}

\item{polarity}{If \code{pattern} is not used, either \code{+} or \code{-}; this chooses which files to look for.}

\item{pattern}{The file pattern to search for (i.e. what is appended to the raw file name).
If a raw file \code{FILENAME.mzXML} was picked into \code{FILENAME.mzXML.MSlist.ABCDE.RData}, the \code{pattern} must be set to \code{ABCDE}, see above.
If this is longer than one, a separate container will be created for each pattern and filled accordingly.}

\item{rtrange}{Optional RT range of peaks to extract. Use this to cut off e.g. the column dead volume.}
}
\value{
A profile container in enviMass format (for no \code{pattern} or a single \code{pattern}, or a list of containers named after the
corresponding pattern/scan.
}
\description{
Reads a set of files containing enviPick results, and fills an enviMass profile container with the data.
Does not compute the actual profiles; this can be done with \link{computeProfiles}.
}
\details{
Code is copied and adapted from enviMass.
}
\author{
stravsmi, mostly copied from Martin Loos
}
