% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screen.R
\name{generateSampleList}
\alias{generateSampleList}
\title{Generate sample table from filenames}
\usage{
generateSampleList(files, sep = "_", names = NULL)
}
\arguments{
\item{files}{Files vector}

\item{sep}{Separator}

\item{names}{Names for output columns. Otherwise will be named col1, col2...}
}
\value{
A data frame with results
}
\description{
Splits a vector of filenames by a separator to make a corresponding sample table.
E.g. 001_file1_15_22.mzXML gets split to 001 - file1 - 15 - 22.
Should handle missing entries more or less
}
\author{
stravsmi
}
