% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ramclustRdf.R
\name{makeProfileInfo}
\alias{makeProfileInfo}
\title{Extract essential info from profiles}
\usage{
makeProfileInfo(profiles, index = 1, name = "")
}
\arguments{
\item{profiles}{A single enviMass profile container or a list of profile containers}

\item{index}{The run index to append}
}
\value{
A data frame with profile_ID, RT, mz, intensity and scantype index.
}
\description{
Extracts minimal essential information from profiles table for each profile,
and adds profile run index as required
}
\author{
stravsmi
}
