% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ramclustRdf.R
\name{makeProfileMatrix}
\alias{makeProfileMatrix}
\title{Make matrix for use with ramclustR.df}
\usage{
makeProfileMatrix(profiles, type = c("ramclust", "graphviz"),
  addIndex = TRUE)
}
\arguments{
\item{profiles}{A profiles object from enviMass}

\item{type}{\code{ramclust} or \code{graphviz}, in the latter case the retention time is inserted as a first row (not only in the column names).}

\item{addIndex}{If \code{TRUE}, add a sample index column in front of the table, as required by RAMClust.}
}
\value{
A matrix m x n, where the rows (sic) are the samples and the columns are the profiles (features) without "profile 0".
The columns are named profileID_RT. This is a requirement for ramclustR (which usually uses mz_RT, but actually doesn't use
mz for calculation and therefore the content is arbitrary. It's easier to use the profile ID for precise lookup.)
}
\description{
Make matrix for use with ramclustR.df
}
