% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screen.R
\name{ppm}
\alias{ppm}
\title{Calculate ppm distance from a mass.}
\usage{
ppm(mz, ppmlimit)
}
\arguments{
\item{mz}{mass}

\item{ppmlimit}{ppm deviation to calculate}
}
\description{
Calculate ppm distance from a mass.
}
\note{
This conflicts with the RMassBank declaration of \code{ppm}, which does something different!
}
\author{
stravsmi
}
