% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ramclustRdf.R
\name{processRc}
\alias{processRc}
\title{Process RAMClust output for viewer}
\usage{
processRc(RC, profile)
}
\arguments{
\item{RC}{Result from RAMClust call}
}
\value{
a dataframe
}
\description{
Process RAMClust output for viewer
}

