% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ramclustRdf.R
\name{processSpectra}
\alias{processSpectra}
\title{Process RAMClust output for viewer}
\usage{
processSpectra(cluster, profileInfo, type = c("ramclust", "graphvis"))
}
\arguments{
\item{RC}{Result from RAMClust call}
}
\value{
a dataframe
}
\description{
Process RAMClust output for viewer
}
