% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ramclustRdf.R
\name{quickclust}
\alias{quickclust}
\title{Quick RAMClustR application}
\usage{
quickclust(profileID, profiles, rttol = 60, raw = FALSE,
  profileMatrix = NULL, sr = 0.5, st = 25)
}
\arguments{
\item{profileID}{The profile you want to look up (integer)}

\item{profiles}{A profile list produced by \link{nontarget}}

\item{rttol}{Retention time window to extract from the profile matrix to cluster on. A value of 60 means +- 30 
(in unit of your retention time, i.e. if you have seconds use seconds)}

\item{raw}{if TRUE, return the RAMClustR result instead. Useful for development purposes.}

\item{profileMatrix}{A sample x feature matrix as returned by \link{\code{makeProfileMatrix}}}
}
\value{
a \code{data.frame} (\code{profileID, mz, RT, int}) with the "pseudospectrum" at that RT 
       (using the pseudo-sum-intensity from RAMClustR)
}
\description{
Applies RAMClustR to a small subset of all profiles that are in the retention time window around a specific profile of interest.
}
