% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readMSdata.direct.R
\name{readMSdata.direct}
\alias{readMSdata.direct}
\title{readMSdata.direct}
\usage{
readMSdata.direct(mz1, MSlevel = c(1), progbar = FALSE, minRT = FALSE,
  maxRT = FALSE, minmz = FALSE, maxmz = FALSE)
}
\description{
Creates a MSlist object directly from a readMzXmlFile list (i.e. allows for manipulating the list beforehand)
Otherwise c/p from enviPick::readMSdata()
}

