% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{readScantype}
\alias{readScantype}
\title{Convert scantype string to readable output}
\usage{
readScantype(x)
}
\arguments{
\item{x}{input characters vector}
}
\value{
Converted vector
}
\description{
Convert scantype string to readable output
}
\author{
stravsmi
}
