% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings.R
\name{RmsSettingsTemplate}
\alias{RmsSettingsTemplate}
\title{Create RMassScreening settings template file}
\usage{
RmsSettingsTemplate(target = "settings.ini")
}
\arguments{
\item{target}{Target filename}
}
\description{
Create RMassScreening settings template file
}

