% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{runViewer}
\alias{runViewer}
\title{Run the results viewer}
\usage{
runViewer(totalTable, hits, timepoints, sampleGroups, patterns = NULL,
  clusters = NULL, addcols = 1, settings = getOption("RMassScreening"),
  profiles = NULL, files = NULL, ...)
}
\arguments{
\item{totalTable}{}

\item{hits}{}

\item{timepoints}{}

\item{sampleGroups}{}

\item{patterns}{}

\item{clusters}{Optional RAMClust output converted into suitable input by \code{\link{rcProcess}}}

\item{addcols}{Number of additional columns per sample group present in summarized (totalTable) output. This is 1 if regular processing (adding the mean) is used.}

\item{...}{}
}
\description{
Run the results viewer
}
