% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/timetrends.R
\name{sampleSelector}
\alias{sampleSelector}
\title{Generalized sample selector which hopefully works automatically like it is supposed to.
Takes a sampleList and a sampleAssignment.}
\usage{
sampleSelector(sampleList, sampleAssignment, ...)
}
\description{
The sampleList is the table merged to the filename list via assignSamples, and can contain any arbitrary columns
describing "what" the measured sample is. E.g.
filename, sample, timepoint, workup
(where "workup" can be different ways of working up a biological sample, e.g. lysed cells vs supernatant)
}
\details{
The sampleAssignment table is the table building relationships between samples, and can contain arbitrary columns
describing what group a sample belongs to.
E.g.
sample, organism, spikedChemicals, stressor

sampleSelector returns the indices of all samples from sampleList which match all specified conditions (be they in
sampleList or sampleAssignemnt). E.g.

sampleSelector(l, a, workup = "supernatant", spikedChemicals = "atrazine")
sampleSelector(l, a, species = c("ecoli", "botulinum"), spikedChemicals = "atrazine")
}

