% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ramclustRdf.R
\name{selectSpectraProfiles}
\alias{selectSpectraProfiles}
\title{Select spectra for profile container}
\usage{
selectSpectraProfiles(w, scantype)
}
\arguments{
\item{w}{msmsWorkspace}

\item{scantype}{Scantype (profile container) to select}
}
\value{
msmsWorkspace
}
\description{
From an msmsWorkspace containing DDA or DIA spectra, select (in index table) those for a specific profile container for use in the viewer.
}
\author{
stravsmi
}
