% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ramclustRdf.R
\name{submatrix}
\alias{submatrix}
\title{Make RAMClust input submatrix for a retention time window}
\usage{
submatrix(rt, profiles, rttol = 60, addIndex = TRUE)
}
\arguments{
\item{rt}{Center retention time}

\item{profiles}{profile container}

\item{rttol}{Retention time tolerance}

\item{addIndex}{Add sample index in front of table? Needed for RAMClust}
}
\value{
Matrix
}
\description{
Creates a profile matrix (RAMClust input) from a subset of a profile container (enviMass) in a retention time slice
}
\author{
stravsmi
}
