% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/screen.R
\name{assignSamples}
\alias{assignSamples}
\title{"files" must be the files list in the same order they were used when building profiles.
"sampleList" needs a filename column
The filename from sampleList is coordinated with a "sampleIDs" ID from profiles via the "files" list,
and the rest of the sampleList table is merged into it.}
\usage{
assignSamples(files, sampleList)
}
\description{
"files" must be the files list in the same order they were used when building profiles.
"sampleList" needs a filename column
The filename from sampleList is coordinated with a "sampleIDs" ID from profiles via the "files" list,
and the rest of the sampleList table is merged into it.
}

