% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/masses.R
\name{combineReactions}
\alias{combineReactions}
\title{Expects two data frames with columns mass, name.
Note that the "reactions" dataframe will have dash-separated names,
such as "oh-deme"}
\usage{
combineReactions(substances, reactions, sep = " ", omit.name = "")
}
\description{
Expects two data frames with columns mass, name.
Note that the "reactions" dataframe will have dash-separated names,
such as "oh-deme"
}

