% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ramclustRdf.R
\name{makeProfileMatrix}
\alias{makeProfileMatrix}
\title{Make matrix for use with ramclustR.df}
\usage{
makeProfileMatrix(profiles)
}
\arguments{
\item{profiles}{A profiles object from enviMass}
}
\value{
A matrix m x n, where the rows (sic) are the samples and the columns are the profiles (features) without "profile 0".
The columns are named profileID_RT. This is a requirement for ramclustR (which usually uses mz_RT, but actually doesn't use
mz for calculation and therefore the content is arbitrary. It's easier to use the profile ID for precise lookup.)
}
\description{
Make matrix for use with ramclustR.df
}

