% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{runViewer}
\alias{runViewer}
\title{Run the results viewer}
\usage{
runViewer(totalTable, hits, timepoints, sampleGroups, patterns = NULL,
  clusters = NULL, addcols = 1, settings = getOption("RMassScreening"),
  ...)
}
\arguments{
\item{totalTable}{}

\item{hits}{}

\item{timepoints}{}

\item{sampleGroups}{}

\item{patterns}{}

\item{clusters}{RAMClust output converted into suitable input by \code{\link{rcProcess}}}

\item{addcols}{}

\item{...}{}
}
\description{
Run the results viewer
}

