% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screen.R
\name{screenProfiles}
\alias{screenProfiles}
\title{"profiles" is a profiles container
"suspects" is a suspect list and must contain a "mass" column. Additionally a "name" column is useful
(though not formally required)}
\usage{
screenProfiles(profiles, suspects, polarity = "+", ppmLimit = 2)
}
\description{
"profiles" is a profiles container
"suspects" is a suspect list and must contain a "mass" column. Additionally a "name" column is useful
(though not formally required)
}

